Use `usize` in `Span` instead of `u32`
Even though it will take twice more space on 64-bit systems, it is more
convenient to preserve consistency for now.

+++ syn/src/cursor.rs
@@ -18,8 +18,8 @@ impl<'a> Cursor<'a> {
     }
 
     pub fn span(self) -> Span {
-        let start = self.offset as u32;
-        let end = (self.offset + self.remaining.len()) as u32;
+        let start = self.offset;
+        let end = self.offset + self.remaining.len();
 
         Span::new(start, end)
     }
